Renamed files to a proper camelCase names
